qa_crowbarsetup.sh: Fix Test Updates after IBS maintenance switch
There is no longer a consolidated update channel available on
you.suse.de. Instead we need to have two update channels, one
with the released updates and one additional with the unreleased
ones.

This solution is not perfect since it doesn't work in the
crowbar-register case but at least the autoyast installed nodes
are getting test updates aga.